requirements.txt
ace_tools==0.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.0
fsspec==2025.3.2
gensim==4.3.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gower==0.1.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hdbscan==0.8.40
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
openai==1.93.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==4.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
